Device:
Arm Subsystem base: Remove build command and set the layer in the calling class.
Instruments:
Enforce setting the dispose option when creating the instrument panel form.

Fixes reset known state to use property setter so as to raise the property changed event.
K2000:
Fixes the code for setting the system auto range state.
Fixes the code for setting the trigger source and direction.
K7000:
Fixes the code for setting the trigger source
Fixes the code for setting the trigger source and direction.